+++ src/main/java/ch/ethz/matsim/av/routing/interaction/AVInteractionFinder.java
@@ -1,0 +1,9 @@
+package ch.ethz.matsim.av.routing.interaction;
+
+import org.matsim.facilities.Facility;
+
+public interface AVInteractionFinder {
+	Facility findPickupFacility(Facility fromFacility, double departureTime);
+
+	Facility findDropoffFacility(Facility toFacility, double departureTime);
+}